Commit No 4 - Custom Insector
Started creating the custom insecptor for the region script. Currently shows each point within the poly but with a bug that makes the floatfields spread apart making inputting into the Z feild impossible.

Fix bug
Add buttons to add and remove from list
Add button to return  to a default state
Add button to debug a random point and spawn a marker on it for testing.