Commit No 5 - Correcting Version and Custom Inspector
Changed the Unity version to match the one in College
Made progress of custom inspector fixing the bug of not being able to chagne vert transforms within the inspector.
Added inspector buttons for adding, removing and setting from and to defaults.

Next step, finish the custom inspector and find new way to get the prefab region for instaiation as pathway reference can be easily broken.