Commit No 6 - Region Creation
Previous method of creating the regions through menu items used a project path to get a prefab. This could be broken easily from someone moving the file or not having it in the correct place making the project path incorrect. Now the method creates the objects instead of just instaiated making the need for a prefab null.