Commit No 7 - Region Creation part 2
The previous method of creating the gameobject used a static list as a default for verts, this caused many problems. This method removes that list and adds the default spawning of a region to the main region class list.

Next steps is to add the buttons for setting a default and loading that default