Commit No 8 - Presets
Changed from default functionality to a dictionary that holds presets. UI of custom inspecotor is fully set up but functionlaity is still buggy.

Next step, fix the bug where the saved preset is linked to the region that made the save meaning that changing that region changes the preset. This also makes any region that uses that preset become linked to the othe region making any changes to a region happen to the other as well.