Commit No 14 - Demo Scene
Created a Demo Scene.
Made a basic AI and destination cube to show the desitination of the AI and it walking toward it,.
Come across a bug to do with the region manager being set to null when the game comes out of play mode but fixed the bug with a new variable setting function.

Possible next steps:
Adding a test function to the region to allow 50 cubes to spawn using the pickrandomdestination function to test the region.
Adding holes to the region, will require a large amount of refactoring code to work with a new script.
Adding A new function with an additional set of getting the terrain pos for Y to allow for non-navmeshagent AI to function.
Adding Flight-regions, a region identical to the current but with additional information of min-max height which will also be randomly chosen to given 3D vectors.