Commit No 17 - Backtrack
Backtracking progress due to the notice of a very problematic bug. The variables of the editor script get reset when entering playmode. The previous method of solving only works if the user clicks on a region before adding a new region, if they don't the new region will also create a new region manager.