Commit No 18 - Fixes and Demo scenes
Fixes the lose of refernce to region manager when entering and exiting playmode causing newly created regions to create a second region manager
Created the raycast demoscene
Tested the raycast function and found/solve bugs
Currently the Raycast function requires a layer int instead of layermask and the demo scene requires the terrain layer to be pre configured. I want to change this if possible.