Commit No 21 - starting raycast/flight testing
The start of the functions within the editor to test the region if its a raycat or flight region.
Also starts another round of refactoring changing editor script to be more userfriendly with tooltips and warnings.
Next steps will be to finish implimenting the testing features, refactoring the editor script completely with tooltips, warnings and better naming conventions. I also want to change the next of MaxTerrainHeight and other variables to be better suited to what they actually are. In terms of MaxTerrainHeight it is actually the RaycastHeight and doens't even need to be remotely close to the terrrains max height.