Commit No 24.2 - Preset data saving setup
Makes the classes, and started the data class, for saving the preset data. Is more complex than I thought so waiting for another day to properly plan out.

DECISION TO MAKE:
either use a List of a List of float arrays.
attempt a Vector3 serialisation surragate to allow for a list of vector3s to simplify the saving process but add the creation of this addition class.