Commit 26 - Reexport and User Experience
Exported an up-to-date version of the tool to a Unity package alongside putting it in a more noticable file for when users downlaod the whole project from GitHub.
Added additional user experince method to the saving and loading of preset data since there isn't functions to automate this easily adding in warning errors to the inspector when there is unloaded data and unsaved data will make it more unlikely for user data to be lost.