Commit 13 - Drag and Drop, Item Databases
Added the 2nd to last feature of drag and drop, removing items from inventory when dragging to item off the inventory. This brings up a box to how many you wan't to destroy.
Created the Item Database scriptable object with functions for finding items based on ID and the automatic populate of database through a list.